@@ -36,18 +36,12 @@
     "vue-router": "^4.0.15"
   },
   "devDependencies": {
+    "@antfu/eslint-config": "^0.25.0",
     "@originjs/vite-plugin-commonjs": "^1.0.1",
     "cross-spawn": "^5.0.1",
     "echarts": "^5.3.2",
-    "eslint": "^3.19.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-html": "^3.0.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.2.0",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-standard": "^3.0.1",
+    "eslint": "^8.16.0",
+    "eslint-plugin-html": "^6.2.0",
     "highlight.js": "^9.12.0",
     "jest": "^21.2.0",
     "jest-serializer-vue": "^0.3.0",
@@ -60,6 +54,7 @@
     "selenium-server": "^3.0.1",
     "shelljs": "^0.7.6",
     "stylus": "^0.56.0",
+    "typescript": "^4.7.2",
     "vite": "^2.7.2",
     "vite-plugin-compression": "^0.4.0",
     "vite-plugin-env-compatible": "^1.1.1",
