@@ -1,10 +1,4 @@
-<template>
-  <div class="discuss-wrap">
-    <div id="uyan_frame"></div>
-  </div>
-</template>
 <script>
-
 export default {
   mounted () {
     const node = document.createElement('script')
@@ -12,15 +6,22 @@ export default {
     const frame = document.querySelector('.discuss-wrap')
     frame.appendChild(node)
   },
-  destroyed () {
+  unmounted () {
     // 友言根据这些变量得知是否已经加载评论
     // 删除这些变量会使友言重新加载
     delete window.UYAN
     delete window.UYAN_L
     delete window.uyan_loadover
     delete window.uyan_loaded
-  }
+  },
 }
 </script>
+
+<template>
+  <div class="discuss-wrap">
+    <div id="uyan_frame" />
+  </div>
+</template>
+
 <style lang="stylus">
 </style>
