@@ -1,30 +1,12 @@
-<template>
-  <div>
-    <Row type="flex" justify="start">
-      <Col :span="2" class="label">Title</Col>
-      <Col :span="21">
-        <Input v-model="news.title"></Input>
-      </Col>
-    </Row>
-    <Row>
-      <Col :span="23">
-        <vue-editor id="editor1"
-          useCustomImageHandler
-          @imageAdded="handleImageAdded" v-model="news.content">
-        </vue-editor>
-      </Col>
-    </Row>
-  </div>
-</template>
 <script>
-import api from '@/api'
 import { VueEditor } from 'vue2-editor'
-import { useNewsStore } from '@/store/modules/news'
 import { mapState } from 'pinia'
+import api from '@/api'
+import { useNewsStore } from '@/store/modules/news'
 
 export default {
   computed: {
-    ...mapState(useNewsStore, ['news'])
+    ...mapState(useNewsStore, [ 'news' ]),
   },
   methods: {
     handleImageAdded (file, Editor, cursorLocation) {
@@ -35,14 +17,37 @@ export default {
           const url = data.url // Get url from response
           Editor.insertEmbed(cursorLocation, 'image', url)
         })
-        .catch((err) => console.log(err))
-    }
+        .catch(err => console.log(err))
+    },
   },
   components: {
-    VueEditor
-  }
+    VueEditor,
+  },
 }
 </script>
+
+<template>
+  <div>
+    <Row type="flex" justify="start">
+      <Col :span="2" class="label">
+        Title
+      </Col>
+      <Col :span="21">
+        <Input v-model="news.title" />
+      </Col>
+    </Row>
+    <Row>
+      <Col :span="23">
+        <VueEditor
+          id="editor1"
+          v-model="news.content"
+          use-custom-image-handler @imageAdded="handleImageAdded"
+        />
+      </Col>
+    </Row>
+  </div>
+</template>
+
 <style lang="stylus" scoped>
 .ivu-row-flex
   margin-bottom: 20px
