@@ -1,3 +1,27 @@
+<script>
+import { VueEditor } from 'vue2-editor'
+import api from '@/api'
+
+export default {
+  components: {
+    VueEditor,
+  },
+  props: [ 'problem' ],
+  methods: {
+    handleImageAdded (file, Editor, cursorLocation) {
+      const formData = new window.FormData()
+      formData.append('image', file)
+      api.getImage(formData)
+        .then(({ data }) => {
+          const url = data.url // Get url from response
+          Editor.insertEmbed(cursorLocation, 'image', url)
+        })
+        .catch(err => console.log(err))
+    },
+  },
+}
+</script>
+
 <template>
   <div class="proadd-wrap">
     <Row>
@@ -21,70 +45,61 @@
         </Input>
       </Col>
     </Row>
-    <div class="label">Description</div>
+    <div class="label">
+      Description
+    </div>
     <Row>
       <Col :span="23">
-        <vue-editor id="editor1"
-          useCustomImageHandler
-          @imageAdded="handleImageAdded" v-model="problem.description">
-        </vue-editor>
+        <VueEditor
+          id="editor1"
+          v-model="problem.description"
+          use-custom-image-handler @imageAdded="handleImageAdded"
+        />
       </Col>
     </Row>
-    <div class="label">Input</div>
+    <div class="label">
+      Input
+    </div>
     <Row>
       <Col :span="23">
-        <vue-editor id="editor2" v-model="problem.input"></vue-editor>
+        <VueEditor id="editor2" v-model="problem.input" />
       </Col>
     </Row>
-    <div class="label">Output</div>
+    <div class="label">
+      Output
+    </div>
     <Row>
       <Col :span="23">
-        <vue-editor id="editor3" v-model="problem.output"></vue-editor>
+        <VueEditor id="editor3" v-model="problem.output" />
       </Col>
     </Row>
-    <div class="label">Hint</div>
+    <div class="label">
+      Hint
+    </div>
     <Row>
       <Col :span="23">
-        <vue-editor id="editor4" v-model="problem.hint"></vue-editor>
+        <VueEditor id="editor4" v-model="problem.hint" />
       </Col>
     </Row>
-    <div class="label">Sample Input</div>
+    <div class="label">
+      Sample Input
+    </div>
     <Row>
       <Col :span="23">
-        <Input type="textarea" :rows="8" v-model="problem.in"></Input>
+        <Input v-model="problem.in" type="textarea" :rows="8" />
       </Col>
     </Row>
-    <div class="label">Sample Output</div>
+    <div class="label">
+      Sample Output
+    </div>
     <Row>
       <Col :span="23">
-        <Input type="textarea" :rows="8" v-model="problem.out"></Input>
+        <Input v-model="problem.out" type="textarea" :rows="8" />
       </Col>
     </Row>
   </div>
 </template>
-<script>
-import api from '@/api'
-import { VueEditor } from 'vue2-editor'
 
-export default {
-  props: ['problem'],
-  methods: {
-    handleImageAdded (file, Editor, cursorLocation) {
-      const formData = new window.FormData()
-      formData.append('image', file)
-      api.getImage(formData)
-        .then(({ data }) => {
-          const url = data.url // Get url from response
-          Editor.insertEmbed(cursorLocation, 'image', url)
-        })
-        .catch((err) => console.log(err))
-    }
-  },
-  components: {
-    VueEditor
-  }
-}
-</script>
 <style lang="stylus">
 .proadd-wrap
   margin-bottom: 20px
