@@ -1,13 +1,13 @@
 // The Vue build version to load with the `import` command
 // (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import {createApp} from 'vue'
+import { createApp } from 'vue'
+import ViewUIPlus from 'view-ui-plus'
+import { createPinia } from 'pinia'
 import App from './App'
 import router from './router'
 // import VueClipboard from 'vue-clipboard2'
-import ViewUIPlus from 'view-ui-plus'
 import '@/my-theme/index.less'
 import { semiRestful } from './api'
-import { createPinia } from 'pinia'
 import { useSessionStore } from '@/store/modules/session'
 import { useRootStore } from '@/store'
 
@@ -19,9 +19,14 @@ const app = createApp(App)
 app.use(pinia)
 app.use(ViewUIPlus)
 
+// https://forum.vuejs.org/t/how-to-use-globalproperties-in-vue-3-setup-method/108387/4
+// https://vuejs.org/api/application.html#app-provide
+app.provide('$Message', app.config.globalProperties.$Message)
+app.provide('$Modal', app.config.globalProperties.$Modal)
+
 Promise.all([
   useSessionStore().fetch(),
-  useRootStore().fetchWebsiteConfig()
+  useRootStore().fetchWebsiteConfig(),
 ]).then(() => {
   if (useRootStore().semi_restful) {
     semiRestful()
@@ -30,8 +35,6 @@ Promise.all([
   app.use(router)
   // https://www.mathew-paul.nz/posts/how-to-use-vue2-with-vite/
   app.mount('#app')
-  console.log(app.component.$Message)
-  console.log(app._component)
 })
 
 // remove @vue/composition-api when migrated to Vue 3
