@@ -1,20 +1,20 @@
-import api from '@/api'
 import { defineStore } from 'pinia'
+import api from '@/api'
 
 export const useDiscussStore = defineStore('discuss', {
   state: () => ({
     list: [],
-    discuss: {}
+    discuss: {},
   }),
   actions: {
     async findOne (payload) {
-      const {data} = await api.discuss.findOne(payload)
+      const { data } = await api.discuss.findOne(payload)
       data.discuss.comments = data.discuss.comments.sort((x, y) => x.create - y.create)
       this.discuss = data.discuss
       return data
     },
     async find (payload) {
-      const {data} = await api.discuss.find(payload)
+      const { data } = await api.discuss.find(payload)
       if (data.list != null && Array.isArray(data.list)) {
         data.list = data.list.sort((x, y) => -x.update + y.update)
       }
@@ -28,7 +28,7 @@ export const useDiscussStore = defineStore('discuss', {
     },
     async delete (payload) {
       await api.discuss.delete(payload)
-      this.list = this.list.filter((p) => p.did !== +(payload.did))
-    }
-  }
+      this.list = this.list.filter(p => p.did !== +(payload.did))
+    },
+  },
 })
