@@ -1,20 +1,20 @@
-import api from '@/api'
 import { defineStore } from 'pinia'
+import api from '@/api'
 
 export const useSolutionStore = defineStore('solution', {
   state: () => ({
     list: [],
     solution: null,
-    sum: 0
+    sum: 0,
   }),
   actions: {
     async find (payload) {
-      const {data} = await api.solution.find(payload)
+      const { data } = await api.solution.find(payload)
       this.list = data.list.docs
       this.sum = data.list.total
     },
     async findOne (payload) {
-      const {data} = await api.solution.findOne(payload)
+      const { data } = await api.solution.findOne(payload)
       this.solution = data.solution
     },
     create (payload) {
@@ -25,6 +25,6 @@ export const useSolutionStore = defineStore('solution', {
     },
     clearCode () {
       this.solution.code = ''
-    }
-  }
+    },
+  },
 })
