@@ -1,6 +1,6 @@
+import { defineStore } from 'pinia'
 import api from '@/api'
 import constant from '@/util/constant'
-import { defineStore } from 'pinia'
 
 export const useStatisticsStore = defineStore('statistics', {
   state: () => ({
@@ -8,26 +8,26 @@ export const useStatisticsStore = defineStore('statistics', {
     countList: [],
     sumCharts: 0,
     table: [],
-    sumStatis: 0
+    sumStatis: 0,
   }),
   actions: {
     async find (payload) {
-      const {data} = await api.getStatistics(payload)
+      const { data } = await api.getStatistics(payload)
       this.list = data.list
       this.countList = data.countList
       this.sumCharts = data.sumCharts
       this.sumStatis = data.sumStatis
-      let obj = {
+      const obj = {
         name: 'Total Submissions',
-        num: data.sumStatis
+        num: data.sumStatis,
       }
       this.table.push(obj)
       for (let i = 0; i < constant.statisTableObj.length; i++) {
-        let obk = {}
+        const obk = {}
         obk.name = constant.statisTableObj[i]
         obk.num = data.countList[i]
         this.table.push(obk)
       }
-    }
-  }
+    },
+  },
 })
