@@ -1,18 +1,18 @@
-import api from '@/api'
 import { defineStore } from 'pinia'
+import api from '@/api'
 
 export const useTestcaseStore = defineStore('testcase', {
   state: () => ({
     list: [],
-    testcase: {}
+    testcase: {},
   }),
   actions: {
     async findOne (payload) {
-      const {data} = await api.testcase.findOne(payload)
+      const { data } = await api.testcase.findOne(payload)
       this.testcase = data
     },
     async find (payload) {
-      const {data} = await api.testcase.find(payload)
+      const { data } = await api.testcase.find(payload)
       this.list = data
       return data
     },
@@ -20,8 +20,8 @@ export const useTestcaseStore = defineStore('testcase', {
       return api.testcase.create(payload)
     },
     async delete (payload) {
-      const {data} = await api.testcase.delete(payload)
+      const { data } = await api.testcase.delete(payload)
       this.list = data
-    }
-  }
+    },
+  },
 })
