@@ -1,8 +1,8 @@
 import moment from 'moment'
 
 function formate (value) {
-  let num = value * 100
-  return num.toFixed(2) + '%'
+  const num = value * 100
+  return `${num.toFixed(2)}%`
 }
 
 function timePretty (time) {
@@ -14,17 +14,17 @@ function timeagoPretty (time) {
 }
 
 function timeContest (time) {
-  let h = Math.floor(time / (1000 * 60 * 60))
-  let m = Math.floor((time - h * 1000 * 60 * 60) / (1000 * 60))
-  let s = Math.floor((time - h * 1000 * 60 * 60 - m * 1000 * 60) / 1000)
-  let ss = ('0' + s).split('').reverse().join('').substr(0, 2).split('').reverse().join('')
-  let mm = ('0' + m).split('').reverse().join('').substr(0, 2).split('').reverse().join('')
-  return h + ':' + mm + ':' + ss
+  const h = Math.floor(time / (1000 * 60 * 60))
+  const m = Math.floor((time - h * 1000 * 60 * 60) / (1000 * 60))
+  const s = Math.floor((time - h * 1000 * 60 * 60 - m * 1000 * 60) / 1000)
+  const ss = (`0${s}`).split('').reverse().join('').substr(0, 2).split('').reverse().join('')
+  const mm = (`0${m}`).split('').reverse().join('').substr(0, 2).split('').reverse().join('')
+  return `${h}:${mm}:${ss}`
 }
 
 export {
   formate,
   timePretty,
   timeContest,
-  timeagoPretty
+  timeagoPretty,
 }
