@@ -1,3 +1,18 @@
+<script>
+export default {
+  methods: {
+    home () {
+      this.$router.push({
+        name: 'home',
+      })
+    },
+    back () {
+      this.$router.go(-1)
+    },
+  },
+}
+</script>
+
 <template>
   <div>
     <Card>
@@ -8,30 +23,21 @@
       </Row>
       <Row type="flex" justify="center">
         <Col :offset="8" :span="4">
-          <Button type="primary" size="large" @click="home">Home</Button>
+          <Button type="primary" size="large" @click="home">
+            Home
+          </Button>
         </Col>
         <Col :span="4">
-          <Button type="primary" size="large" @click="back">Back</Button>
+          <Button type="primary" size="large" @click="back">
+            Back
+          </Button>
         </Col>
         <Col :span="8" />
       </Row>
     </Card>
   </div>
 </template>
-<script>
-export default {
-  methods: {
-    home () {
-      this.$router.push({
-        name: 'home'
-      })
-    },
-    back () {
-      this.$router.go(-1)
-    }
-  }
-}
-</script>
+
 <style lang="stylus" scoped>
 p
   font-size: 80px
