@@ -1,30 +1,23 @@
-<template>
-  <div>
-    <h1>新增消息</h1>
-    <oj-news-edit></oj-news-edit>
-    <Button type="primary" size="large" @click="submit">Submit</Button>
-  </div>
-</template>
 <script>
-import NewsEdit from '@/components/NewsEdit'
 import { mapActions, mapState } from 'pinia'
+import NewsEdit from '@/components/NewsEdit'
 import { useNewsStore } from '@/store/modules/news'
 
 export default {
   data: () => ({
     addNews: {
       title: '',
-      content: ''
-    }
+      content: '',
+    },
   }),
   computed: {
-    ...mapState(useNewsStore, ['news'])
+    ...mapState(useNewsStore, [ 'news' ]),
   },
   created () {
     useNewsStore().setCurrentNews(this.addNews)
   },
   methods: {
-    ...mapActions(useNewsStore, ['create']),
+    ...mapActions(useNewsStore, [ 'create' ]),
     submit () {
       if (!this.news.title) {
         this.$Message.error('Title can not be empty')
@@ -35,13 +28,24 @@ export default {
             this.$router.push({ name: 'newsInfo', params: { nid } })
           })
       }
-    }
+    },
   },
   components: {
-    'oj-news-edit': NewsEdit
-  }
+    OjNewsEdit: NewsEdit,
+  },
 }
 </script>
+
+<template>
+  <div>
+    <h1>新增消息</h1>
+    <OjNewsEdit />
+    <Button type="primary" size="large" @click="submit">
+      Submit
+    </Button>
+  </div>
+</template>
+
 <style lang="stylus" scoped>
 h1
   margin-bottom: 20px
