@@ -1,60 +1,8 @@
-<template>
-  <div>
-    <h1>管理用户组</h1>
-    <Row type="flex" justify="start">
-      <Col :span="2" class="label">Group</Col>
-      <Col :span="4">
-        <Select v-model="ind" filterable>
-          <Option v-for="(item, index) in groupList" :value="index" :key="item.gid">{{ item.title }}</Option>
-        </Select>
-      </Col>
-      <Col :offset="1" :span="2">
-        <Dropdown @on-click="manageGroup">
-          <Button type="primary">
-            Manage
-            <Icon type="ios-arrow-down"></Icon>
-          </Button>
-          <template #list>
-            <DropdownMenu>
-              <DropdownItem name="search">Search</DropdownItem>
-              <DropdownItem name="create">Create</DropdownItem>
-              <DropdownItem name="delete">Delete</DropdownItem>
-            </DropdownMenu>
-          </template>
-        </Dropdown>
-      </Col>
-      <Col span="2">
-        <Tag>
-          {{ operation }}
-        </Tag>
-      </Col>
-    </Row>
-    <Row type="flex" justify="start">
-      <Col :span="2" class="label">Title</Col>
-      <Col :span="4">
-        <Input v-model="group.title"></Input>
-      </Col>
-    </Row>
-    <Transfer
-      :data="transData"
-      :target-keys="targetKeys"
-      :render-format="format"
-      :list-style="listStyle"
-      :operations="['To left','To right']"
-      filterable
-      :filter-method="filterMethod"
-      @on-change="handleChange"
-      class="tranfer">
-    </Transfer>
-    <Button type="primary" @click="saveGroup" class="submit">Submit</Button>
-  </div>
-</template>
-
 <script>
 import { mapActions, mapState } from 'pinia'
+import only from 'only'
 import { useUserStore } from '@/store/modules/user'
 import { useGroupStore } from '@/store/modules/group'
-import only from 'only'
 
 export default {
   data: () => ({
@@ -62,24 +10,24 @@ export default {
     targetKeys: [],
     listStyle: {
       width: '350px',
-      height: '400px'
+      height: '400px',
     },
     userList: [],
-    operation: 'search'
+    operation: 'search',
   }),
   computed: {
-    ...mapState(useGroupStore, ['group']),
-    ...mapState(useGroupStore, {groupList: 'list'}),
-    ...mapState(useUserStore, ['user']),
+    ...mapState(useGroupStore, [ 'group' ]),
+    ...mapState(useGroupStore, { groupList: 'list' }),
+    ...mapState(useUserStore, [ 'user' ]),
     ...mapState(useUserStore, {
-      userSum: 'list'
+      userSum: 'list',
     }),
     transData () {
       return this.userSum.map((item, index) => ({
-        key: index + '',
-        label: item.uid + ' | ' + item.nick
+        key: `${index}`,
+        label: `${item.uid} | ${item.nick}`,
       }))
-    }
+    },
   },
   created () {
     this.fetchGroup()
@@ -89,9 +37,9 @@ export default {
     this.clearSavedUsers()
   },
   methods: {
-    ...mapActions(useGroupStore, ['find', 'findOne', 'update', 'create', 'clearSavedGroups']),
-    ...mapActions(useGroupStore, {remove: 'delete'}),
-    ...mapActions(useUserStore, ['clearSavedUsers']),
+    ...mapActions(useGroupStore, [ 'find', 'findOne', 'update', 'create', 'clearSavedGroups' ]),
+    ...mapActions(useGroupStore, { remove: 'delete' }),
+    ...mapActions(useUserStore, [ 'clearSavedUsers' ]),
     async fetchGroup () {
       this.$Spin.show()
       await useUserStore().find()
@@ -105,7 +53,7 @@ export default {
       return item.label
     },
     filterMethod (data, query) {
-      return data.label.indexOf(query) > -1
+      return data.label.includes(query)
     },
     handleChange (newTargetKeys) {
       this.targetKeys = newTargetKeys
@@ -122,7 +70,7 @@ export default {
         try {
           await this.findOne({ gid: this.group.gid })
           this.group.list.forEach((item) => {
-            this.targetKeys.push(this.userList.indexOf(item) + '')
+            this.targetKeys.push(`${this.userList.indexOf(item)}`)
           })
         } finally {
           this.$Spin.hide()
@@ -150,16 +98,16 @@ export default {
             },
             onCancel: () => {
               this.$Message.info('已取消删除！')
-            }
+            },
           })
         }
       }
     },
     async saveGroup () {
-      const user = this.targetKeys.map((item) => this.userList[+item])
+      const user = this.targetKeys.map(item => this.userList[+item])
       const group = Object.assign(
         only(this.group, 'gid title'),
-        { list: user }
+        { list: user },
       )
       if (this.group.gid !== '') {
         this.$Spin.showLoading()
@@ -178,10 +126,76 @@ export default {
           this.$Spin.hide()
         }
       }
-    }
-  }
+    },
+  },
 }
 </script>
+
+<template>
+  <div>
+    <h1>管理用户组</h1>
+    <Row type="flex" justify="start">
+      <Col :span="2" class="label">
+        Group
+      </Col>
+      <Col :span="4">
+        <Select v-model="ind" filterable>
+          <Option v-for="(item, index) in groupList" :key="item.gid" :value="index">
+            {{ item.title }}
+          </Option>
+        </Select>
+      </Col>
+      <Col :offset="1" :span="2">
+        <Dropdown @on-click="manageGroup">
+          <Button type="primary">
+            Manage
+            <Icon type="ios-arrow-down" />
+          </Button>
+          <template #list>
+            <DropdownMenu>
+              <DropdownItem name="search">
+                Search
+              </DropdownItem>
+              <DropdownItem name="create">
+                Create
+              </DropdownItem>
+              <DropdownItem name="delete">
+                Delete
+              </DropdownItem>
+            </DropdownMenu>
+          </template>
+        </Dropdown>
+      </Col>
+      <Col span="2">
+        <Tag>
+          {{ operation }}
+        </Tag>
+      </Col>
+    </Row>
+    <Row type="flex" justify="start">
+      <Col :span="2" class="label">
+        Title
+      </Col>
+      <Col :span="4">
+        <Input v-model="group.title" />
+      </Col>
+    </Row>
+    <Transfer
+      :data="transData"
+      :target-keys="targetKeys"
+      :render-format="format"
+      :list-style="listStyle"
+      :operations="['To left', 'To right']"
+      filterable
+      :filter-method="filterMethod"
+      class="tranfer"
+      @on-change="handleChange"
+    />
+    <Button type="primary" class="submit" @click="saveGroup">
+      Submit
+    </Button>
+  </div>
+</template>
 
 <style lang="stylus" scoped>
 .ivu-col-offset-1
