@@ -1,60 +1,8 @@
-<template>
-  <div>
-    <h1>管理标签组</h1>
-    <Row type="flex" justify="start">
-      <Col :span="2" class="label">Tag</Col>
-      <Col :span="4">
-        <Select v-model="ind" filterable>
-          <Option v-for="(item, index) in tagList" :value="index" :key="item.tid">{{ item.tid }}</Option>
-        </Select>
-      </Col>
-      <Col :offset="1" :span="2">
-        <Dropdown @on-click="manageTag">
-          <Button type="primary">
-            Manage
-            <Icon type="ios-arrow-down"></Icon>
-          </Button>
-          <template #list>
-            <DropdownMenu>
-              <DropdownItem name="search">Search</DropdownItem>
-              <DropdownItem name="create">Create</DropdownItem>
-              <DropdownItem name="delete">Delete</DropdownItem>
-            </DropdownMenu>
-          </template>
-        </Dropdown>
-      </Col>
-      <Col span="2">
-        <Tag>
-          {{ operation }}
-        </Tag>
-      </Col>
-    </Row>
-    <Row type="flex" justify="start">
-      <Col :span="2" class="label">Title</Col>
-      <Col :span="4">
-        <Input v-model="tag.tid"></Input>
-      </Col>
-    </Row>
-    <Transfer
-      :data="transData"
-      :target-keys="targetKeys"
-      :render-format="format"
-      :list-style="listStyle"
-      :operations="['To left','To right']"
-      filterable
-      :filter-method="filterMethod"
-      @on-change="handleChange"
-      class="tranfer">
-    </Transfer>
-    <Button type="primary" @click="saveTag" class="submit">Submit</Button>
-  </div>
-</template>
-
 <script>
 import only from 'only'
+import { mapActions, mapState } from 'pinia'
 import { useProblemStore } from '@/store/modules/problem'
 import { useTagStore } from '@/store/modules/tag'
-import { mapActions, mapState } from 'pinia'
 
 export default {
   data: () => ({
@@ -62,26 +10,26 @@ export default {
     targetKeys: [],
     listStyle: {
       width: '350px',
-      height: '400px'
+      height: '400px',
     },
     problemList: [],
     operation: 'search',
-    isNew: false
+    isNew: false,
   }),
   computed: {
     ...mapState(useProblemStore, {
-      problemSum: 'list'
+      problemSum: 'list',
     }),
     ...mapState(useTagStore, {
       tagList: 'list',
-      tag: 'tag'
+      tag: 'tag',
     }),
     transData () {
       return this.problemSum.map((item, index) => ({
-        key: index + '',
-        label: item.pid + ' | ' + item.title
+        key: `${index}`,
+        label: `${item.pid} | ${item.title}`,
       }))
-    }
+    },
   },
   created () {
     this.fetchTag()
@@ -92,11 +40,11 @@ export default {
   },
   methods: {
     ...mapActions(useProblemStore, {
-      findProblems: 'find'
+      findProblems: 'find',
     }),
-    ...mapActions(useProblemStore, ['clearSavedProblems']),
-    ...mapActions(useTagStore, ['find', 'findOne', 'update', 'create', 'clearSavedTags']),
-    ...mapActions(useTagStore, {remove: 'delete'}),
+    ...mapActions(useProblemStore, [ 'clearSavedProblems' ]),
+    ...mapActions(useTagStore, [ 'find', 'findOne', 'update', 'create', 'clearSavedTags' ]),
+    ...mapActions(useTagStore, { remove: 'delete' }),
     async fetchTag () {
       this.$Spin.show()
       const opt = { page: -1 }
@@ -111,7 +59,7 @@ export default {
       return item.label
     },
     filterMethod (data, query) {
-      return data.label.indexOf(query) > -1
+      return data.label.includes(query)
     },
     handleChange (newTargetKeys) {
       this.targetKeys = newTargetKeys
@@ -128,7 +76,7 @@ export default {
         try {
           await this.findOne({ tid: this.tag.tid })
           this.tag.list.forEach((item) => {
-            this.targetKeys.push(this.problemList.indexOf(item) + '')
+            this.targetKeys.push(`${this.problemList.indexOf(item)}`)
           })
         } finally {
           this.$Spin.hide()
@@ -156,16 +104,16 @@ export default {
             },
             onCancel: () => {
               this.$Message.info('已取消删除！')
-            }
+            },
           })
         }
       }
     },
     saveTag () {
-      const problems = this.targetKeys.map((item) => this.problemList[+item])
+      const problems = this.targetKeys.map(item => this.problemList[+item])
       const tag = Object.assign(
         only(this.tag, 'tid'),
-        { list: problems }
+        { list: problems },
       )
       if (!this.isNew) {
         this.$Spin.showLoading()
@@ -185,10 +133,76 @@ export default {
           this.$Spin.hide()
         })
       }
-    }
-  }
+    },
+  },
 }
 </script>
+
+<template>
+  <div>
+    <h1>管理标签组</h1>
+    <Row type="flex" justify="start">
+      <Col :span="2" class="label">
+        Tag
+      </Col>
+      <Col :span="4">
+        <Select v-model="ind" filterable>
+          <Option v-for="(item, index) in tagList" :key="item.tid" :value="index">
+            {{ item.tid }}
+          </Option>
+        </Select>
+      </Col>
+      <Col :offset="1" :span="2">
+        <Dropdown @on-click="manageTag">
+          <Button type="primary">
+            Manage
+            <Icon type="ios-arrow-down" />
+          </Button>
+          <template #list>
+            <DropdownMenu>
+              <DropdownItem name="search">
+                Search
+              </DropdownItem>
+              <DropdownItem name="create">
+                Create
+              </DropdownItem>
+              <DropdownItem name="delete">
+                Delete
+              </DropdownItem>
+            </DropdownMenu>
+          </template>
+        </Dropdown>
+      </Col>
+      <Col span="2">
+        <Tag>
+          {{ operation }}
+        </Tag>
+      </Col>
+    </Row>
+    <Row type="flex" justify="start">
+      <Col :span="2" class="label">
+        Title
+      </Col>
+      <Col :span="4">
+        <Input v-model="tag.tid" />
+      </Col>
+    </Row>
+    <Transfer
+      :data="transData"
+      :target-keys="targetKeys"
+      :render-format="format"
+      :list-style="listStyle"
+      :operations="['To left', 'To right']"
+      filterable
+      :filter-method="filterMethod"
+      class="tranfer"
+      @on-change="handleChange"
+    />
+    <Button type="primary" class="submit" @click="saveTag">
+      Submit
+    </Button>
+  </div>
+</template>
 
 <style lang="stylus" scoped>
 .ivu-col-offset-1
