@@ -1,21 +1,14 @@
-<template>
-  <div>
-    <h1>修改消息</h1>
-    <oj-news-edit></oj-news-edit>
-    <Button type="primary" size="large" @click="submit">Submit</Button>
-  </div>
-</template>
 <script>
-import NewsEdit from '@/components/NewsEdit'
 import { mapActions, mapState } from 'pinia'
+import NewsEdit from '@/components/NewsEdit'
 import { useNewsStore } from '@/store/modules/news'
 
 export default {
   computed: {
-    ...mapState(useNewsStore, ['news'])
+    ...mapState(useNewsStore, [ 'news' ]),
   },
   methods: {
-    ...mapActions(useNewsStore, ['update']),
+    ...mapActions(useNewsStore, [ 'update' ]),
     submit () {
       if (!this.news.title) {
         this.$Message.error('Title can not be empty')
@@ -26,13 +19,24 @@ export default {
             this.$router.push({ name: 'newsInfo', params: { nid } })
           })
       }
-    }
+    },
   },
   components: {
-    'oj-news-edit': NewsEdit
-  }
+    OjNewsEdit: NewsEdit,
+  },
 }
 </script>
+
+<template>
+  <div>
+    <h1>修改消息</h1>
+    <OjNewsEdit />
+    <Button type="primary" size="large" @click="submit">
+      Submit
+    </Button>
+  </div>
+</template>
+
 <style lang="stylus" scoped>
 h1
   margin-bottom: 20px
