@@ -1,19 +1,12 @@
-<template>
-  <div class="news">
-    <h1>{{ news.title }}</h1>
-    <h4>{{ timePretty(news.create) }}</h4>
-    <div class="news-cont" v-html="news.content"></div>
-  </div>
-</template>
 <script>
-import { mapState, mapActions } from 'pinia'
+import { mapActions, mapState } from 'pinia'
 import { useNewsStore } from '@/store/modules/news'
 import { useRootStore } from '@/store'
 import { timePretty } from '@/util/formate'
 
 export default {
   computed: {
-    ...mapState(useNewsStore, ['news'])
+    ...mapState(useNewsStore, [ 'news' ]),
   },
   created () {
     this.findOne(this.$route.params).then(() => {
@@ -21,12 +14,21 @@ export default {
     })
   },
   methods: {
-    ...mapActions(useRootStore, ['changeDomTitle']),
-    ...mapActions(useNewsStore, ['findOne']),
-    timePretty
-  }
+    ...mapActions(useRootStore, [ 'changeDomTitle' ]),
+    ...mapActions(useNewsStore, [ 'findOne' ]),
+    timePretty,
+  },
 }
 </script>
+
+<template>
+  <div class="news">
+    <h1>{{ news.title }}</h1>
+    <h4>{{ timePretty(news.create) }}</h4>
+    <div class="news-cont" v-html="news.content" />
+  </div>
+</template>
+
 <style lang="stylus" scoped>
 h1, h4
   margin-bottom: 10px
